Add proxied canister call spec
This PR adds a possible implementation of proxied canister calls.
It is intended to be implemented by IDPs that provide a front-end for
end users. The goal is to have a generic specification that works for
existing canister calls (given the canister can be extended to incorporate
the consent canister interface).

Possible next steps:
  * Discuss with working group
  * Security review
  * Concrete end to end example