ICRC-25: Remove app and wallet metadata
This PR removes app and wallet metadata from the permission request.
Reason: the currently gives no way to verify the supplied information.
Any way of verifying the app or wallet metadata is dependent on the transport used.
For this reason, the exchange of appMetadata should be handled by the
transport standard.