ICRC-25: Display warning on new RP, address sec finding F01
This PR adds additional guidance to ICRC-25 that signers must
inform users about _new_ relying parties. If possible they should
assist users with the choice of whether to trust this new relying party.

This PR addresses finding F01 of the ICRC-25 security review.

In addition, if fixes a few typos.