Add Signer Standards Overview (#57)
* Add Signer Standards Overview

This PR adds an overview document to put in context the various
ICRCs drafts that are being worked on in this working group. It also
shows how the standards relate to each other and aims to explain (some of)
the design decisions behind them.

* Add motivation section

* Add links to IC spec

* Implement review feedback

Lots of typos fixed and wording improved.

* Add explanation to trust assumptions section

This PR adds an explanation and an example to the trust assumptions
section. Also, the standards-diagram has been changed to make the
channels better recognizable.