ICRC-25: Remove network selection
This PR simplifies the standard by removing the option to
select networks. There is currently little benefit in having
that feature, but it adds complexity to the standard.

If the network selection becomes useful later on, we can still
add it either as an extension or by publishing a new version of
the standard.