ICRC-25: Drop version and simplify errors (#140)
This PR implements the decision taken in #128 for ICRC-25.
This means that all `version` fields are dropped. This has an effect
on errors as they used to include the version in the code.
This is now simplified to just indicate the broader category with the first
digit and the rest being incremental numbers for each error.
The "unknown error" is dropped in favor of a "generic error" with a
description to avoid unhelpful errors.


By dropping the `version` field, some requests have no `params` anymore.
To not confuse readers by writing "Request: none", the heading is changed
to "Request Params", to better align with the JSON RPC spec and to make it
clear that there still is a request being sent.
For consistency, all "Response" headings are changed to "Result".