ICRC-32: Simplify result and avoid duplication of delegation field name (#154)
This simplifys the result by flattening the outermost object. This object
was initially created because the method used to be able to provide multiple
signatures (which is no longer the cause to avoid complexity).

In addition, the outermost `delegation` field is renamed to `signer_delegation`.
This is consistent with `sender_delegation` in the canister call envelope and
avoids having multiple `delegation` fields on different hierarchy levels.