Update draft status according to WG decision (#189)
* Update draft status according to WG decision

In accordance with the decision taken in the last session
the status of the following ICRCs is updated:
* 21
* 25
* 27
* 28
* 29
* 49

Going forward, the working group will try to avoid any breaking changes
to these standards. Until the standards have been tried and tested in
reference implementations, they will remain in that state. Or, if issues
are found, they will be moved back to draft state in order to resolve
them.

* Update Topic of ICRC-27