ICRC-25: Rename Network Error Category to Transport Channel
The "Network" error category is renamed to "Transport Channel" to include
any error related to communications with signer (and the signer with the
IC).

A new error is introduced (4001) to communicate unexpected interruption
of the communication channel between the rp and the signer. In particular
this error can be used if the user dismisses the UI elements of the signer
early without completing the interaction (i.e. close browser tab or
extension window).